elevate and retract do not shut off motors
After thinking about this, it seemed dangerous to shut down the motors when the limit switches were hit.

The motors will shut off (capability built into talons) when the switches are hit,
but they should restart when the switches are no longer active.  Not optimal, but safer than letting the climber plummet.